tst/requirements.txt
websocket-client>=0.32.0
